requirements.txt
asgiref==3.3.4
Django==3.2.3
django-sequences==2.6
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
pyqt-darktheme==1.2.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
python-codicefiscale==0.3.7
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
QDarkStyle==3.0.2
qt-material==2.8.8
QtPy==1.9.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
